extracted fav btn as component
